pyproject.toml
[tool.poetry]
name = "gmbot"
version = "0.4.2"
description = "Telegram bot GMBOT"
authors = ["Nickolsky <romb.intu@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiosignal = "1.3.1"
attrs = "22.2.0"
beautifulsoup4 = "4.11.1"
charset-normalizer = "2.1.1"
frozenlist = "1.3.3"
greenlet = "2.0.1"
idna = "3.4"
multidict = "6.0.4"
netifaces = "0.11.0"
python-dotenv = "0.21.1"
requests = "2.28.2"
soupsieve = "2.3.2.post1"
sqlalchemy = "1.4.46"
sqlalchemy-utils = "0.39.0"
urllib3 = "1.26.14"
yarl = "1.8.2"
alembic = "^1.13.1"
aiogram = "^3.3.0"
loguru = "^0.7.2"
aioschedule = "^0.5.2"
pillow = "^10.2.0"
pytest = "^7.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
